# frontend-react/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Folio API",
    "description": "Folio \u2014 \u667a\u80fd\u8cc7\u7522\u914d\u7f6e",
    "version": "2.0.0"
  },
  "paths": {
...
    "/ticker": {
      "post": {
        "summary": "Add a stock to the watchlist",
        "description": "\u65b0\u589e\u80a1\u7968\u5230\u8ffd\u8e64\u6e05\u55ae\u3002",
        "operationId": "create_ticker_route_ticker_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TickerCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/gurus": {
      "get": {
        "tags": [
          "smart-money"
        ],
        "summary": "List all active gurus",
        "description": "\u53d6\u5f97\u6240\u6709\u555f\u7528\u4e2d\u7684\u5927\u5e2b\u6e05\u55ae\uff08\u542b is_default \u6a19\u8a18\uff09\u3002",
        "operationId": "get_gurus_gurus_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GuruResponse"
                  },
                  "title": "Response Get Gurus Gurus Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "smart-money"
        ],
        "summary": "Add a custom guru by CIK",
        "description": "\u65b0\u589e\u81ea\u8a02\u5927\u5e2b\u3002\u82e5 CIK \u5df2\u5b58\u5728\u4e14\u70ba\u505c\u7528\u72c0\u614b\u5247\u91cd\u65b0\u555f\u7528\u3002",
        "operationId": "create_guru_gurus_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuruCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuruResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gurus/{guru_id}": {
      "delete": {
        "tags": [
          "smart-money"
        ],
        "summary": "Deactivate a guru (soft delete)",
        "description": "\u505c\u7528\u5927\u5e2b\uff08\u8edf\u522a\u9664\uff0c\u4fdd\u7559\u6b77\u53f2\u7533\u5831\u8cc7\u6599\uff09\u3002",
        "operationId": "delete_guru_gurus__guru_id__delete",
        "parameters": [
          {
            "name": "guru_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Guru Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Guru Gurus  Guru Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gurus/sync": {
      "post": {
        "tags": [
          "smart-money"
        ],
        "summary": "Trigger 13F sync for all tracked gurus",
        "description": "\u89f8\u767c\u6240\u6709\u555f\u7528\u4e2d\u5927\u5e2b\u7684 13F \u540c\u6b65\uff08\u540c\u6b65\u57f7\u884c\uff0c\u9069\u5408\u6392\u7a0b\u547c\u53eb\uff09\u3002",
        "operationId": "sync_all_gurus_sync_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncAllResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gurus/{guru_id}/sync": {
      "post": {
        "tags": [
          "smart-money"
        ],
        "summary": "Trigger 13F sync for one guru",
        "description": "\u89f8\u767c\u6307\u5b9a\u5927\u5e2b\u7684 13F \u540c\u6b65\u3002",
        "operationId": "sync_one_gurus__guru_id__sync_post",
        "parameters": [
          {
            "name": "guru_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Guru Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gurus/grand-portfolio": {
      "get": {
        "tags": [
          "smart-money"
        ],
        "summary": "Aggregated portfolio across all active gurus' latest 13F filings",
        "description": "\u8de8\u6240\u6709\u555f\u7528\u4e2d\u5927\u5e2b\u7684\u6700\u65b0 13F \u6301\u5009\u805a\u5408\u8996\u5716\u3002",
        "operationId": "get_grand_portfolio_endpoint_gurus_grand_portfolio_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrandPortfolioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gurus/{guru_id}/filing": {
      "get": {
        "tags": [
          "smart-money"
        ],
        "summary": "Latest 13F filing summary for a guru",
        "description": "\u53d6\u5f97\u6307\u5b9a\u5927\u5e2b\u6700\u65b0\u7533\u5831\u6458\u8981\uff0c\u542b\u5206\u985e\u6301\u5009\uff08\u65b0\u5efa\u5009\u3001\u6e05\u5009\u3001\u52a0\u78bc\u3001\u6e1b\u78bc\uff09\u3002",
        "operationId": "get_filing_gurus__guru_id__filing_get",
        "parameters": [
          {
            "name": "guru_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Guru Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuruFilingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gurus/{guru_id}/holdings": {
      "get": {
        "tags": [
          "smart-money"
        ],
        "summary": "Holding changes for a guru (action != UNCHANGED), sorted by significance",
        "description": "\u53d6\u5f97\u6307\u5b9a\u5927\u5e2b\u6700\u65b0\u7533\u5831\u4e2d\u6709\u52d5\u4f5c\u7684\u6301\u5009\uff08\u6392\u9664 UNCHANGED\uff09\u3002\n\n\u7d50\u679c\u4f9d\u8b8a\u52d5\u5e45\u5ea6 abs(change_pct) \u964d\u5e8f\u6392\u5217\uff0c\u518d\u4f9d weight_pct \u964d\u5e8f\uff0c\u6700\u591a\u56de\u50b3 limit \u7b46\u3002",
        "operationId": "get_holdings_gurus__guru_id__holdings_get",
        "parameters": [
          {
            "name": "guru_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Guru Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Max number of changes to return (sorted by abs(change_pct) then weight_pct)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Max number of changes to return (sorted by abs(change_pct) then weight_pct)"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GuruHoldingResponse"
                  },
                  "title": "Response Get Holdings Gurus  Guru Id  Holdings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gurus/{guru_id}/top": {
      "get": {
        "tags": [
          "smart-money"
        ],
        "summary": "Top N holdings by weight for a guru",
        "description": "\u53d6\u5f97\u6307\u5b9a\u5927\u5e2b\u6301\u5009\u6b0a\u91cd\u6700\u9ad8\u7684\u524d N \u652f\u80a1\u7968\uff08\u9810\u8a2d Top 10\uff0c\u6700\u591a 50\uff09\u3002",
        "operationId": "get_top_holdings_gurus__guru_id__top_get",
        "parameters": [
          {
            "name": "guru_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Guru Id"
            }
          },
          {
            "name": "n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Top N holdings",
              "default": 10,
              "title": "N"
            },
            "description": "Top N holdings"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GuruHoldingResponse"
                  },
                  "title": "Response Get Top Holdings Gurus  Guru Id  Top Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SyncAllResponse": {
        "properties": {
          "synced": {
            "type": "integer",
            "title": "Synced"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "errors": {
            "type": "integer",
            "title": "Errors"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SyncResponse"
            },
            "type": "array",
            "title": "Results",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "synced",
          "skipped",
          "errors"
        ],
        "title": "SyncAllResponse",
        "description": "POST /gurus/sync \u56de\u50b3\u7684\u5168\u5927\u5e2b\u540c\u6b65\u7d50\u679c\u3002"
      },
...
      "TickerCreateRequest": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "category": {
            "$ref": "#/components/schemas/StockCategory"
          },
          "thesis": {
            "type": "string",
            "title": "Thesis"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "is_etf": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Etf"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "category",
          "thesis"
        ],
        "title": "TickerCreateRequest",
        "description": "POST /ticker \u8acb\u6c42 Body\u3002"
      },
...
    }
  }
}